+++ docs/apps/setmeasurement/index.rst
@@ -1,0 +1,18 @@
+Set Measurement
+===============
+
+.. versionadded:: 0.6.0
+
+The set measurement CLI applications provide a simple way to conduct monitoring
+measurements to a set of predefined target points, using a GeoCom capable
+total station.
+
+.. note::
+
+    The various subcommands are available as CLI submodules.
+
+.. toctree::
+    :maxdepth: 1
+
+    setup
+    measure
